refactor: OauthType Member 도메인의 enum으로 리팩토링
