--- src/main/java/com/ssafy/ssafsound/domain/auth/domain/OauthType.java
@@ -1,15 +1,0 @@
-package com.ssafy.ssafsound.domain.auth.domain;
-
-public enum OauthType {
-    GOOGLE("google"), GITHUB("github"), KAKAO("kakao"), APPLE("apple");
-
-    private final String oauthName;
-
-    OauthType(String oauthName) {
-        this.oauthName = oauthName;
-    }
-
-    public boolean isEqual(String oauthName) {
-        return this.oauthName.equals(oauthName);
-    }
-}
